requirements.txt
playwright 
beautifulsoup4 
pydantic 
typer 
streamlit 
httpx 
python-dotenv 
pytest

